+++ examples/user.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -34,8 +34,13 @@ options = {
     'mobilePhone': '555-2121',
     'officePhone': '555-1212',
     'password': 'myPass1234',
-    'roles': ['Read only'],
-    'type': 'UserAndRoles',
+    'permissions': [
+        {
+            'roleName': 'Read only',
+            'scopeUri': '/rest/scopes/b7259c1c-c52a-4be3-82de-d50e1730695e',
+        }
+    ],
+    'type': 'UserAndPermissions',
     'userName': 'testUser'
 }
 
@@ -43,41 +48,36 @@ options = {
 config = try_load_from_file(config)
 
 oneview_client = OneViewClient(config)
+users = oneview_client.users
 
 # Create a User
-user = oneview_client.users.create(options)
-print("Created user '%s' successfully.\n  uri = '%s'\n" % (user['userName'], user['uri']))
+user = users.create(options)
+print("Created user '%s' successfully.\n  uri = '%s'\n" % (user.data['userName'], user.data['uri']))
 
-# Adds different roles to the recently created user
-user = oneview_client.users.get_by('userName', 'testUser')
-print(user)
-print("\n user: %s, roles before update: %s" % (user['userName'], user['roles']))
-user['replaceRoles'] = True
-user['roles'] = ['Infrastructure administrator']
-user = oneview_client.users.update(user)
-print("\n user: %s, roles after update: %s" % (user['userName'], user['roles']))
-
-# Get user by role
-user = oneview_client.users.get_by('role', 'Read only')
-print("Found users by role: '%s'.\n '\n" % (user))
+# Change Password
+change_password_request = {
+    "currentPassword": "admin12345",
+    "enabled": "true",
+    "password": "admin1234",
+    "userName": "admin"
+}
+changePasswordResponse = users.change_password(change_password_request)
+print("Changed Password successfully")
+print(changePasswordResponse)
 
 # Get user by name
-user = oneview_client.users.get_by('userName', 'testUser')
-print("Found user by name: '%s'. uri = '%s'\n" % (user['userName'], user['uri']))
+user = users.get_by_userName(options['userName'])
+print("Found user by uri = '%s'\n" % user.data['uri'])
 
 # Get all users
 print("Get all users")
-users = oneview_client.users.get_all()
+users = users.get_all()
 pprint(users)
 
 # # Validates if full name is already in use
-bol = oneview_client.users.validate_full_name(options['fullName'])
+bol = users.validate_full_name(options['fullName'])
 print("Is full name already in use? %s" % (bol))
 
 # # Validates if user name is already in use
-bol = oneview_client.users.validate_user_name(options['userName'])
+bol = users.validate_user_name(options['userName'])
 print("Is user name already in use? %s" % (bol))
-
-# Delete the created user
-oneview_client.users.delete(user)
-print("\nSuccessfully deleted user")